Refactor

+++ MM/oracles/simple_prices.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Callable
 import requests
 
@@ -8,29 +9,29 @@ SOURCE_DATA = {
     3: 'https://data-api.binance.vision/api/v3/aggTrades?symbol=BTCUSDC',
 }
 
-def fetch_price_from_url(url: str) -> float:
+def fetch_price_from_url(url: str) -> Decimal:
     response = requests.get(url, timeout=5)
     response.raise_for_status()
     data = response.json()
     if not data:
         raise ValueError("No data returned from API")
-    return float(sorted(data, key = lambda x: x['T'])[-1]['p'])  
+    return Decimal(sorted(data, key = lambda x: x['T'])[-1]['p'])  
 
 # Default processors for simple markets that don't need to process price 
 # in any way
-def simple_fetcher(url: str) -> Callable[[], float]:
+def simple_fetcher(url: str) -> Callable[[], Decimal]:
     return lambda: fetch_price_from_url(url)
 
 
-def wbtc_dog_price_fetcher() -> float:
+def wbtc_dog_price_fetcher() -> Decimal:
     btc_price = fetch_price_from_url('https://data-api.binance.vision/api/v3/aggTrades?symbol=BTCUSDC')
     dog_price = requests.get('https://api.gateio.ws/api/v4/spot/trades?currency_pair=DOG_USDT&limit=1')
-    dog_price = float(dog_price.json()[0]['price'])
+    dog_price = Decimal(dog_price.json()[0]['price'])
 
     return btc_price / dog_price
 
 
-PRICE_FETCHERS: dict[int, Callable[[], float]] = {
+PRICE_FETCHERS: dict[int, Callable[[], Decimal]] = {
     1: simple_fetcher(SOURCE_DATA[1]),
     2: simple_fetcher(SOURCE_DATA[2]),
     3: simple_fetcher(SOURCE_DATA[3]),
@@ -38,7 +39,7 @@ PRICE_FETCHERS: dict[int, Callable[[], float]] = {
 }
 
 
-def get_price_fetcher(market_id: int) -> Callable[[], float]:
+def get_price_fetcher(market_id: int) -> Callable[[], Decimal]:
     '''
     Returns a function that is used to fetch price for given `market_id`.
